refactor: 유효성 검사 로직 관리하는 폴더 추가
