+++ src/domain/buyLottoTickets.js
@@ -1,24 +1,11 @@
 import { PRICE_PER_LOTTO } from "./constants.js";
 import LottoTicket from "./LottoTicket.js";
+import { validatePurchaseAmount } from "./validators/purchaseValidator.js";
 
 export const buyLottoTickets = (purchaseAmount) => {
-  if (!isValidPurchaseAmount(purchaseAmount)) {
-    throw new InvalidPurchaseAmount();
-  }
+  validatePurchaseAmount(purchaseAmount);
 
   const lottoCount = Math.trunc(purchaseAmount / PRICE_PER_LOTTO);
 
   return Array.from(Array(lottoCount), () => new LottoTicket());
 };
-
-const isValidPurchaseAmount = (purchaseAmount) => {
-  if (!Number.isInteger(purchaseAmount)) return false;
-  if (purchaseAmount < PRICE_PER_LOTTO) return false;
-  return true;
-};
-
-export class InvalidPurchaseAmount extends Error {
-  constructor() {
-    super(`구입 금액은 ${PRICE_PER_LOTTO}원 이상의 정수여야 합니다.`);
-  }
-}
